More transition to later package versions Remove old navigation in preparation for new navigation/app bar
