+++ src/components/FriendList.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import FriendCard from "./FriendCard";
-import { ImageList } from "@material-ui/core";
+import { ImageList } from "@mui/material";
 
 const FriendList = (props) => {
   return (
