+++ src/components/FriendModal/FriendModal.tsx
@@ -1,6 +1,6 @@
 import React, { useState, useContext, useEffect } from "react";
 import { Box, Button, Typography, Modal } from "@mui/material";
-import { FormControl, IconButton, TextField } from "@material-ui/core";
+import { FormControl, IconButton, TextField } from "@mui/material";
 import SearchIcon from "@material-ui/icons/Search";
 import { useFriendsContext } from "../../context/useFriendsContext";
